Title: 이차원 배열의 오름차순 정리, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250217/이차원 배열의 오름차순 정리/ascending-order-of-two-dimensional-array.cpp
@@ -1,22 +1,22 @@
 #include <iostream>
 using namespace std;
 
-int N, K;
+long long N, K;
 
-bool isValid(int x) {
-    int count = 0;
-    for (int i = 1; i <= N; i++) {
+bool isValid(long long x) {
+    long long count = 0;
+    for (long long i = 1; i <= N; i++) {
         count += min(N, x / i);
     }
     return count >= K;
 }
 
-int binarySearch() {
-    int left = 1, right = N * N;
-    int result = 0;
+long long binarySearch() {
+    long long left = 1, right = N * N;
+    long long result = 0;
 
     while (left <= right) {
-        int mid = (left + right) / 2;
+        long long mid = (left + right) / 2;
 
         if (isValid(mid)) { // mid 값이 K번째 숫자 이상일 경우
             result = mid;   // 가능한 답이므로 저장
